Length method should return int.
FastR does not support long vectors.

+++ com.oracle.truffle.r.runtime/src/com/oracle/truffle/r/runtime/data/altrep/AltRepClassDescriptor.java
@@ -234,8 +234,8 @@ public abstract class AltRepClassDescriptor extends RBaseObject {
         if (logger.isLoggable(Level.FINER)) {
             logAfterInteropExecute(ret);
         }
-        assert ret instanceof Long;
-        return ((Long) ret).intValue();
+        assert ret instanceof Integer;
+        return (int) ret;
     }
 
     @Override
